refactor: extract bounded execution foundation

Copilot-Session: d78e26c7-9fe6-4d34-a8bb-4afa87fc5f84
